+++ gwt-nio/src/main/java/java/nio/DoubleBuffer.java
@@ -91,7 +91,7 @@ public abstract class DoubleBuffer extends Buffer implements Comparable<DoubleBu
      */
     public static DoubleBuffer wrap(double[] array, int start, int len) {
         int length = array.length;
-        if (start < 0 || len < 0 || (long)start + (long)len > length) {
+        if (start < 0 || len < 0 || start + len > length) {
             throw new IndexOutOfBoundsException();
         }
 
@@ -318,7 +318,7 @@ public abstract class DoubleBuffer extends Buffer implements Comparable<DoubleBu
      */
     public DoubleBuffer get(double[] dest, int off, int len) {
         int length = dest.length;
-        if (off < 0 || len < 0 || (long)off + (long)len > length) {
+        if (off < 0 || len < 0 || off + len > length) {
             throw new IndexOutOfBoundsException();
         }
         
@@ -482,7 +482,7 @@ public abstract class DoubleBuffer extends Buffer implements Comparable<DoubleBu
      */
     public DoubleBuffer put(double[] src, int off, int len) {
         int length = src.length;
-        if (off < 0 || len < 0 || (long)off + (long)len > length) {
+        if (off < 0 || len < 0 || off + len > length) {
             throw new IndexOutOfBoundsException();
         }
