+++ gwt-nio/src/main/java/java/nio/ReadWriteFloatArrayBuffer.java
@@ -105,7 +105,7 @@ final class ReadWriteFloatArrayBuffer extends FloatArrayBuffer {
 
     public FloatBuffer put(float[] src, int off, int len) {
         int length = src.length;
-        if (off < 0 || len < 0 || (long)off + (long)len > length) {
+        if (off < 0 || len < 0 || off + len > length) {
             throw new IndexOutOfBoundsException();
         }
         if (len > remaining()) {
