+++ gwt-nio/src/main/java/java/nio/ShortArrayBuffer.java
@@ -65,7 +65,7 @@ abstract class ShortArrayBuffer extends ShortBuffer {
 
     public final ShortBuffer get(short[] dest, int off, int len) {
         int length = dest.length;
-        if (off < 0 || len < 0 || (long)off + (long)len > length) {
+        if (off < 0 || len < 0 || off + len > length) {
             throw new IndexOutOfBoundsException();
         }
         if (len > remaining()) {
