+++ gwt-nio/src/main/java/java/nio/ShortBuffer.java
@@ -92,7 +92,7 @@ public abstract class ShortBuffer extends Buffer implements Comparable<ShortBuff
                 if (array == null) {
                         throw new NullPointerException();
                 }
-                if (start< 0 || len < 0 || (long)start + (long)len > array.length) {
+                if (start< 0 || len < 0 || start + len > array.length) {
                         throw new IndexOutOfBoundsException();
                 }
 
@@ -315,7 +315,7 @@ public abstract class ShortBuffer extends Buffer implements Comparable<ShortBuff
      */
     public ShortBuffer get(short[] dest, int off, int len) {
         int length = dest.length;
-        if (off < 0 || len < 0 || (long)off + (long)len > length) {
+        if (off < 0 || len < 0 || off + len > length) {
             throw new IndexOutOfBoundsException();
         }
         if (len > remaining()) {
@@ -476,7 +476,7 @@ public abstract class ShortBuffer extends Buffer implements Comparable<ShortBuff
      */
     public ShortBuffer put(short[] src, int off, int len) {
         int length = src.length;
-        if (off < 0 || len < 0 || (long)off + (long)len > length) {
+        if (off < 0 || len < 0 || off + len > length) {
             throw new IndexOutOfBoundsException();
         }
